refactor(datastores): decouple store selection from the cli layer
SetDeviceStore previously accepted *cli.Command and read the --datastore flag itself, coupling the persistence package to internal/cli. It now takes a resolved store-type string; flag parsing moves to a new command-layer helper, internal/util/store.Setup, which every call site now uses. pkg/datastores no longer imports internal/cli directly.
