+++ cmd/add/prefix.go
@@ -30,6 +30,7 @@ import (
 	"log"
 
 	"github.com/Cray-HPE/cani/internal/cli"
+	"github.com/Cray-HPE/cani/internal/util/store"
 	"github.com/Cray-HPE/cani/pkg/datastores"
 	"github.com/Cray-HPE/cani/pkg/devicetypes"
 	"github.com/google/uuid"
@@ -62,7 +63,7 @@ Examples:
 func addPrefix(cmd *cli.Command, args []string) error {
 	cidr := args[0]
 
-	if err := datastores.SetDeviceStore(cmd, args); err != nil {
+	if err := store.Setup(cmd); err != nil {
 		return fmt.Errorf("failed to set device store: %w", err)
 	}
