+++ cmd/export/connections.go
@@ -30,6 +30,7 @@ import (
 	"os"
 
 	"github.com/Cray-HPE/cani/internal/cli"
+	"github.com/Cray-HPE/cani/internal/util/store"
 	"github.com/Cray-HPE/cani/pkg/datastores"
 	"github.com/Cray-HPE/cani/pkg/devicetypes/connections"
 	"gopkg.in/yaml.v3"
@@ -54,7 +55,7 @@ Example:
 }
 
 func exportConnections(cmd *cli.Command, args []string) error {
-	if err := datastores.SetDeviceStore(cmd, args); err != nil {
+	if err := store.Setup(cmd); err != nil {
 		return fmt.Errorf("failed to set device store: %w", err)
 	}
