+++ cmd/show/show.go
@@ -32,6 +32,7 @@ import (
 	"strings"
 
 	"github.com/Cray-HPE/cani/internal/cli"
+	"github.com/Cray-HPE/cani/internal/util/store"
 	"github.com/Cray-HPE/cani/pkg/datastores"
 	"github.com/Cray-HPE/cani/pkg/devicetypes"
 	"github.com/Cray-HPE/cani/pkg/visual"
@@ -108,7 +109,7 @@ func loadInventory(cmd *cli.Command, args []string) (*devicetypes.Inventory, err
 		return loadInventoryFromFile(filePath)
 	}
 
-	if err := datastores.SetDeviceStore(cmd, args); err != nil {
+	if err := store.Setup(cmd); err != nil {
 		return nil, fmt.Errorf("failed to set device store: %w", err)
 	}
 	return datastores.Datastore.Load()
